[Fix] #67 - DTO 폴더 나누기
